+++ test/jest/helpers/renderWithReduxForm.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  reduxForm,
   reducer as formReducer,
 } from 'redux-form';
 import {
@@ -10,7 +9,6 @@ import {
 import { Provider } from 'react-redux';
 
 const renderWithReduxForm = (component, initialStateValues = {}, formFieldValues = {}) => {
-  const onSubmit = jest.fn();
   const fieldReducer = (state = initialStateValues) => state;
   const reducer = combineReducers({
     field: fieldReducer,
@@ -18,16 +16,9 @@ const renderWithReduxForm = (component, initialStateValues = {}, formFieldValues
   });
   const store = createStore(reducer);
 
-  const Decorated = reduxForm({
-    form: 'testForm',
-    onSubmit: { onSubmit },
-  })(component);
-
   return (
     <Provider store={store}>
-      <Decorated
-        {...formFieldValues}
-      />
+      {component(formFieldValues)}
     </Provider>
   );
 };