+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change history for ui-tenant-settings
 
-## 10.1.0 (IN PROGRESS)
+## 11.0.0 (IN PROGRESS)
 
 * [UITEN-327](https://folio-org.atlassian.net/browse/UITEN-327) Invalidate queries after updating location
+* [UITEN-330](https://folio-org.atlassian.net/browse/UITEN-330) *BREAKING* Refactor locale settings handling.
 
 ## [10.0.0](https://github.com/folio-org/ui-tenant-settings/tree/v10.0.0)(2025-03-12)